src/requirements.txt
aiomysql==0.2.0
Flask==3.1.0
mysql-connector-python==8.4.0
python_bcrypt==0.3.2
Requests==2.32.3
rich==14.0.0
waitress==3.0.2
